pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=75.3.2"]

[project]
name = "django-select2-forms"
dynamic = ["version"]
description = "Django form fields using the Select2 jQuery plugin"
readme = "README.rst"
license = { text = "BSD-2-Clause" }
requires-python = ">=3"
authors = [
    { name = "The Atlantic", email = "programmers@theatlantic.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "django-sortedm2m",
]

[project.urls]
Homepage = "https://github.com/theatlantic/django-select2-forms"

[tool.setuptools]
zip-safe = false
platforms = ["any"]
include-package-data = true

[tool.setuptools.packages.find]
include = ["select2*"]
namespaces = false

[tool.setuptools.dynamic]
version = { attr = "select2.__version__" }
readme = { file = ["README.rst"] }

[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_test.py"
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = "--tb=short --create-db --cov=select2 --cov-branch --cov-report=xml"
django_find_project = false
testpaths = "tests"
pythonpath = "."
